Perform out-of-tree builds with Makefile
Apart from being highly desirable on its own to avoid clutter in the
root folder, this also fixes an error that occurs during cmake
out-of-tree builds if there are left-overs from a make in-tree build.

After some consideration and research, I believe that there is no
reasonable way in cmake to isolate the unwanted files from the build.
The simplest option is to perform out-of-tree builds in make as well.

In fact, the existence of in-tree generated header files during
out-of-tree builds can always lead to inconsistencies between different
translation units because we cannot reliably enfore that the out-of-tree
header is used from everywhere.

Resolves #7.

Notably, the compile-time error in #7 occured due to the fact that the
include guards of in/out-of-tree headers were different, which could
lead to both variants of a header being included, and hence redefinition
errors. However, this is not the actual issue, but rather a symptom that
we were lucky enough to observe.